added test, fixed code climate issues
